requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
face_recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.76.0
frozenlist==1.3.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
orjson==3.6.8
Pillow==9.1.0
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
retina-face==0.0.10
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
